requirements.txt
aiobotocore==2.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==25.12.0
bleach==6.3.0
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.76.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.25.0
dvc==3.65.0
dvc-data==3.17.0
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.127.0
fastjsonschema==2.21.2
filelock==3.20.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.5
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.1
lark==1.3.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
mlflow==3.8.0
mlflow-skinny==3.8.0
mlflow-tracing==3.8.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.4.0
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pyflakes==3.4.0
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
s3fs==2025.12.0
s3transfer==0.15.0
scikit-learn==1.8.0
scipy==1.16.3
scmrepo==3.6.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.50.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.16.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==1.17.3
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
